+++ src/components/SideBar/MobileSideBar.tsx
@@ -1,9 +1,8 @@
 import { FoldIcon, HamburgerIcon } from '@assets';
 import Popup from '@components/Popup';
 import TodoCreateModal from '@components/TodoModal/TodoCreateModal';
-import useDeleteGoal from '@hooks/api/goalsAPI/useDeleteGoal';
-import { useState } from 'react';
 import MobileSideBarContents from './MobileSideBarContents';
+import useModalControl from './components/useModalControll';
 
 interface MobileSideBarProps {
   isOpen: boolean;
@@ -18,23 +17,16 @@ function MobileSideBar({
   userData,
   goalData,
 }: MobileSideBarProps) {
-  const [showTodoModal, setShowTodoModal] = useState(false);
-  const [isDeletePopupVisible, setIsDeletePopupVisible] = useState(false);
-  const [goalId, setGoalId] = useState<number>(0);
-  const { mutate: deleteGoalMutate } = useDeleteGoal();
-  const onShowTodoModal = () => setShowTodoModal(true);
-  const onShowDeletePopup = (id: number) => {
-    setIsDeletePopupVisible(true);
-    setGoalId(id);
-  };
-  const handleDelete = () => {
-    deleteGoalMutate(goalId, {
-      onSuccess: () => {
-        setIsDeletePopupVisible(false);
-        setGoalId(0);
-      },
-    });
-  };
+  const {
+    showTodoModal,
+    isDeletePopupVisible,
+    onShowTodoModal,
+    onShowDeletePopup,
+    handleDelete,
+    setShowTodoModal,
+    setIsDeletePopupVisible,
+  } = useModalControl();
+
   return (
     <>
       <div className="h-12 w-full bg-white">
